feat: Add dynamic count of contributed exams
This commit adds a dynamic count of the number of contributed exams to the webpage. The count is displayed in a hidden element with the id "count". If the user's last visit was more than 30 minutes ago or if it's their first visit, the count is shown along with a message indicating the continuous update of exams. Otherwise, the count is hidden. Additionally, a fade-out effect is applied to the count after 10 seconds to hide it.